Modified Error Checking Script in 'app.js' to Delete Errors from previous Form Submission Before Issuing New Errors
Modified the previously made script in the 'app.js' file to now delete the children of the error panel before adding in new errors when the form is submitted subsequent to the first time. This will ensure errors in the error panel from previous form submissions are deleted before errors from the current submission are added to the panel. This is done using the Document Object Model (DOM). It increases overall readability of errors for the user and creates for a better experience of the web application.

Signed-off-by: Muhammad Ahmed <moahzed@gmail.com>