Updated Footer Message
Updated footer message, which warns users to read the 'README.md' markdown file for proper documentation and instructions on how to use the app. This is there so that users may know exactly how the form validation app is supposed to function and may test it accordingly. The message now gives greater detail by providing the exact section in the 'README.md' file which holds the instructions.

Signed-off-by: Muhammad Ahmed <moahzed@gmail.com>